+++ psutil/_psutil_sunos.c
@@ -1248,6 +1248,8 @@ void init_psutil_sunos(void)
 #else
     PyObject *module = Py_InitModule("_psutil_sunos", PsutilMethods);
 #endif
+    PyModule_AddIntConstant(module, "version", _PSUTIL_VERSION);
+
     PyModule_AddIntConstant(module, "SSLEEP", SSLEEP);
     PyModule_AddIntConstant(module, "SRUN", SRUN);
     PyModule_AddIntConstant(module, "SZOMB", SZOMB);